Refactor quantity and stock fields to support decimal values across models and APIs
- Updated models (CatalogStockSyncResult, PosMonitorItem, Product, ProductVariety, RefundItemRequest, SalesLineItem, AnalyticsNamedValue, etc.) to change quantity and stock fields from int to double.
- Modified JSON serialization/deserialization to handle double values correctly.
- Adjusted API endpoints (PosApi, InventoryController, OrderController, etc.) to accept and return decimal quantities for stock and quantity fields.
- Added migration to widen integer columns to decimal in relevant tables (order_items, products, etc.) to support fractional quantities.
- Enhanced low stock notifications and payroll reports to reflect changes in quantity handling.

+++ myfinalpos/lib/models/catalog_stock_sync.dart
@@ -8,24 +8,24 @@ class CatalogStockSyncResult {
 
   final String revision;
   final bool unchanged;
-  final Map<String, int> products;
-  final Map<String, int> varieties;
+  final Map<String, double> products;
+  final Map<String, double> varieties;
 
   factory CatalogStockSyncResult.fromJson(Map<String, dynamic> json) {
     final rawProducts = json['products'];
     final rawVarieties = json['varieties'];
-    final products = <String, int>{};
-    final varieties = <String, int>{};
+    final products = <String, double>{};
+    final varieties = <String, double>{};
 
     if (rawProducts is Map) {
       rawProducts.forEach((key, value) {
-        products[key.toString()] = (value as num?)?.toInt() ?? 0;
+        products[key.toString()] = (value as num?)?.toDouble() ?? 0;
       });
     }
 
     if (rawVarieties is Map) {
       rawVarieties.forEach((key, value) {
-        varieties[key.toString()] = (value as num?)?.toInt() ?? 0;
+        varieties[key.toString()] = (value as num?)?.toDouble() ?? 0;
       });
     }
